Pass login credentials to the wopi-controller if available.
Use case are authenticated mounts which otherwise cannot be accessed.

The procedure is:

- obtain a cookie from the client and generate an app-password
- pass that cookie encrypted with the server secret on to the Collabora
  server as URL parameter
- in the WopiController, see if that cookie is there, decrypt it and get
  the login-credentials for the user

This should not make things worse as they are already: the app-passwords
are actually passphrases for login credentials stored in the token
table. So the NC database already stores passwords.

The Wopi-token and cookie are erase latest at logout-time. The cookie
expires at the time when the Wopi-token expires.