Implement authenticated WOPI requests if needed.
If the mount claims to need authentication then an app-token encrypted
with the WOPI token is generated in order to allow for authentication.

This is a security hole. However, passing personalized parameters along
with the wopi-url to the Collabora server breaks everthing. Seemingly
the server distinguishes the document-instances by looking at the WOPI
URL.